move filters to own component, add + more text
