+++ carbonmark/components/pages/Projects/styles.ts
@@ -47,62 +47,6 @@ export const dropdown = css`
   }
 `;
 
-export const pillContainer = css`
-  position: relative;
-  gap: 0.4rem;
-  width: 100%;
-  display: flex;
-  overflow-x: auto;
-  grid-column: full;
-  align-items: center;
-  max-width: 168rem;
-  align-items: center;
-  justify-self: center;
-
-  &::-webkit-scrollbar {
-    display: none;
-  }
-
-  ${breakpoints.desktop} {
-    overflow-x: hidden;
-  }
-
-  &:after {
-    ${breakpoints.desktop} {
-      content: "";
-      top: 0;
-      right: 0;
-      width: 11.5rem;
-      height: 3.2rem;
-      position: absolute;
-      background: linear-gradient(
-        270deg,
-        #eeeff5 55.13%,
-        rgba(238, 239, 245, 0) 100%
-      );
-    }
-  }
-`;
-
-export const pill = css`
-  gap: 0.8rem;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  white-space: nowrap;
-  padding: 0.8rem 1.6rem;
-  background: #8b8fae;
-  border-radius: 4rem;
-  color: white;
-  font-weight: 600;
-  font-size: 1.2rem;
-  line-height: 1.6rem;
-
-  & svg {
-    cursor: pointer;
-  }
-`;
-
 export const projectsList = css`
   grid-column: full;
   justify-self: center;
